+++ src/sources/oscillator/square.rs
@@ -147,6 +147,7 @@ mod tests {
         let config = Config {
             phase_increment: 0.5f32,
             amplitude: 1.0f32,
+            half_threshold: 0.5f32,
         };
 
         let mut oscillator = SquareOscillator::with_config(config);
@@ -171,6 +172,7 @@ mod tests {
         let config = Config {
             phase_increment: 0.25f32,
             amplitude: 2.0f32,
+            half_threshold: 0.5f32,
         };
 
         let mut oscillator = SquareOscillator::with_config(config);
@@ -193,6 +195,7 @@ mod tests {
         let config = Config {
             phase_increment: 0.3f32,
             amplitude: 1.0f32,
+            half_threshold: 0.5f32,
         };
 
         let mut oscillator = SquareOscillator::with_config(config);
@@ -214,6 +217,7 @@ mod tests {
         let config = Config {
             phase_increment: 0.5f64,
             amplitude: 1.0f64,
+            half_threshold: 0.5f64,
         };
         let mut oscillator = SquareOscillator::<f64>::with_config(config);
         let samples: Vec<_> = (0..4).map(|_| oscillator.source().unwrap()).collect();